BUG: Verify that the rawTag is acutally valid before using it.
A crash was occurring when looking up a tag from a sha. This is
because not all tags have a git_tag pointer that is valid. The reason
for this is the tag is not fully instantiated and thus the underlying
library does not create a git_tag for it.